+++ src/WeightedRandomGenerator.java
@@ -2,31 +2,32 @@ import java.util.Arrays;
 import java.util.Random;
 
 /**
- * User: bingjun
+ * @author Bingjun
+ * Generated weighted random numbers according to the weights array.
  */
 
 public class WeightedRandomGenerator {
 
     private Random random = new Random(System.nanoTime());
-    private float[] totals;
+    private double[] totals;
 
-    public WeightedRandomGenerator(float[] weights) {
-        totals = new float[weights.length];
+    public WeightedRandomGenerator(double[] weights) {
+        totals = new double[weights.length];
         initTotals(weights);
     }
 
     public int nextInt() {
-        float sumOfWeights = totals[totals.length - 1];
-        float randomNumber = random.nextFloat() * sumOfWeights;
+        double sumOfWeights = totals[totals.length - 1];
+        double randomNumber = random.nextFloat() * sumOfWeights;
         int index = Arrays.binarySearch(totals, randomNumber);
         index = (index >= 0) ? index : - index - 1;
         return index;
     }
 
-    private void initTotals(float[] weights) {
+    private void initTotals(double[] weights) {
         float runningTotal = 0;
         int i = 0;
-        for (float weight: weights) {
+        for (double weight: weights) {
             runningTotal += weight;
             totals[i++] = runningTotal;
         }
